backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
duckdb==1.1.1
python-multipart==0.0.9
PyJWT==2.9.0
openpyxl>=3.1.0
chardet>=5.0.0
